requirements.txt
altgraph==0.17.4
anyio==4.4.0
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lanms_neo==1.0.2
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnxruntime-gpu==1.16.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pefile==2023.2.7
pillow==10.3.0
Polygon3==3.0.9.1
premailer==3.10.0
protobuf==3.20.2
psutil==6.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.7
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pyreadline3==3.4.1
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qimage2ndarray==1.10.0
rapidfuzz==3.9.3
rarfile==4.2
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
termcolor==2.4.0
tifffile==2024.6.18
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
